Get testOnMessageSendTaskIdMismatch working
We were missing termination of the EventQueue, so the EventConsumer is just hanging.

The Python implementation uses functionality from asyncio.Task,
which I have tried to replicate here, although it is not identical.

Introducing a EventQueueClosedException if trying to read from an empty queue
that has been closed, and use this as the exit condition for the EventConsumer.

Then close the queue after we have called the agent. But since
this is happening in a separate thread wait for the consumer to start listening to
messages before closing. Otherwise the consumer will fail and return immediately.