test: Added unit tests for `RequestContext` (#127)
This PR includes the following changes:  

1. Added a unit test class for the `RequestContext` class, and all unit
tests have been verified to pass. This is based on the unit test class
from the `a2a` Python sdk:
`tests/server/agent_execution/test_context.py`, with a few tests that
are not applicable to the Java version removed.

2. Based on the unit test class, an issue in the `RequestContext` class
was identified and fixed: the `relatedTasks` setting in the constructor
was not working properly.

Fixes #60

---------

Signed-off-by: Sun Yuhan <sunyuhan1998@users.noreply.github.com>
Co-authored-by: Sun Yuhan <sunyuhan1998@users.noreply.github.com>