fix: Fixed the intermittent failure issue in `testOnMessageStreamNewMessageSendPushNotificationSuccess` (#161)
The reason for its intermittent failure, I believe, is as follows:

We initially set up the `CountDownLatch` with a count of 6, and we will
call `latch.countDown()` at the following three locations:

- A. in `onSubscribe`
- B. in the `finally` block of the `post` method in `HttpClient`
- C. in `onNext`

Ideally, when the test method runs, the sequence of `countDown()` calls
should be like this:
**A B C B C B C**

The `countDown()` function was called a total of 7 times!

This explains why we sometimes see only **2 elements** in `results` when
we do `assertEquals(3, results.size())`. It's because the last `onNext`
may not have completed execution yet.

**What this PR does:**  

We could fix this by initializing the `CountDownLatch` to 7. However,
considering the purpose of this test method, I chose instead to keep the
count at 6 and **remove the `latch.countDown()` call in `onSubscribe`**.

Fixes #140

Signed-off-by: Sun Yuhan <sunyuhan1998@users.noreply.github.com>
Co-authored-by: Sun Yuhan <sunyuhan1998@users.noreply.github.com>