chore: Re-organize sub-modules to extract the client part from core (#149)
# Description

This PR is made for better organize the different sub-modules,
essentially the core one by extracing the specification (the model) from
the client, in a similar way as what as been done for the server part.

The goal for this split is to make it easier to provide other client
implementation in a near future.

Although code as been spread across new modules, it does not impact the
way it is used as the same packages as been kept. Therefore, it does not
impact much the main readme

- [X ] Follow the [`CONTRIBUTING` Guide](../CONTRIBUTING.md).
- [ X] Make your Pull Request title in the
<https://www.conventionalcommits.org/> specification.
- Important Prefixes for
[release-please](https://github.com/googleapis/release-please):
- `fix:` which represents bug fixes, and correlates to a
[SemVer](https://semver.org/) patch.
- `feat:` represents a new feature, and correlates to a SemVer minor.
- `feat!:`, or `fix!:`, `refactor!:`, etc., which represent a breaking
change (indicated by the `!`) and will result in a SemVer major.
- [ X] Ensure the tests pass
- [ X] Appropriate READMEs were updated (if necessary)

Fixes #<issue_number_goes_here> 🦕

---------

Co-authored-by: Kabir Khan <kkhan@redhat.com>